requirements.txt
requests
pytest
pyyaml
faker